Remove 'debian' plugin
This plugin provides the alias 'dia' which conflicts with the
diagramming program 'dia'.  On examining the 'debian' plugin, its main
action is to set up a number of debian-related aliases, none of which I
have ever used. So, rather than mess about reversing a single unwanted
alias, simply remove the plugin to avoid the problem entirely.